edid: don't automatically generate edid profiles by default
given that some linux applications are color managed and others
aren't, an auto-assigned auto-generated edid profile will cause
extreme confusion for many users, as two of their applications
which used to display images identically "suddenly" show a
single image differently.

since the user didn't assign/activate the edid profile, there
is no obvious way for a regular user to detect or trace their
issue to xiccd (as many users will be unaware they're running
xiccd or using color management in general).

also keeping in mind that edid profiles never contain a vcgt,
the usefulness of such a profile is quite limited, therefore
it could be considered that auto-assigning auto-generated
profiles to be a bad (but sadly common) practice.

this commit introduces a --edid flag, to restore the old
but arguably broken behavior.